Fix stroke offset, add doc, optimize strokes
I added a lot of documentation for existing functions.
It appears that the offset bug was because I used the wrong bounding
points for the calculation / forgot a parenthesis in the javascript
side.
The stroke optimize is back, but instead of a fix value, I now use the
size of the brush to tell which new stroke points to discard.