Refactor Toolbox to follow the patern in Input
Instead of having a dedicated "ToolMsg" module for messages sent by
toolbox tools, it is now part of the Canvas module, and shares the same
messages as the Input generic type. It also shares the update function
with Input. This reduces redudant code. This also enable me to handle
the logic behind Tool and Input types directly in the canvas module and
exposing those functions rather than having other modules implement it
everytime.