Fix the makefile dependencies, add debug build
make would not compile new changes to the elm codebase if the Index.elm
file was not directly edited. Now if any file in the src folder is
modified, it will try to compile it.

This also adds the ability to build a debug build directly through make.
It was removed in a previous commit.