Refactor canvas code to use relativemouseevent
This commit removes a lot of code that only existed as boilerplate for
abstractions that are unnecessary, I moved all of the "art" and "pen"
code into the Canvas.elm file and deleted a bunch of useless code.
I took the opportunity to rely on an external lib for the mouse movement
capture, this removes completly the need for javascript code and removes
port modules made to interact with it.