Improve pen behavior on edge cases
Literally... When leaving and re-entering the border of the canvas, now
the messages sent to the server are normalized. No warnings should be
raised anymore. I added a new mouse event to handle the case where the
mouse pointer enters the canvas while pressed down. This is a very
common thing to do.

I also sepparated the decoder files from the logic, for readability.