Refactor server
Separate game logic from connection logic. The WebSocketGame module
declares a class that a network handling module can use to plug into
websockets. The GameSocketListener implements WebSocketGame and declares
a ReactiveGame class which a game should implement so it is
"adaptable" into a WebSocketGame.

The trick here is that all the boilerplate connection handling is done
in the implementation of GameSocketListener, this way, a game can be
implemented in a pure way using the ReactiveGame class.

I've also updated the API and documentation to reflect the next
challenges to surrmount. Next commits should focus on implementing
chat.