Port backend to rust
Haskell is great for corectness, however, its drawbacks are especially
fell for this particular project. I needed something that can easily be
ported to ARM and preferably compiled to the raspberry pi without
compiling on the same hardware, I also needed something not as
memory-hungry as Haskell.

Rust has a lot of the qualities of Haskell, and none of the mentioned
limits, this makes it the perfect candidate.

We are still early in the project, this lets me port the software
without too much friction beyond simply learning the new API for the
server framework.

Somehow, I'm still using a server framework called "warp" ;)