Concider any websocket ingress error as closing
Previously, a broken closing handshake would throw the server into an
infinite loop of websocket poll and error propagation. Now the websocket
channel is closed on first irregularity. This fits very nicely with the
architecture, as then a reconnection is forced through the conventional
means. The reconnection tend to happen smoothlya.