Completely remove a closed room from the server
Previously, a game manager would be terminated only if access was
attempted after the last player left. Now the reference to the room
is directly removed from the server state uppon closing.

This prevents a panic that happens when trying to access an empty room,
and also guarentees that the resources are freed even when not accessed
afterward (which was likely)