Simplify ManagerRequest
Now messages that are aimed at the game are all in the same enum
alternative.

While the `loop_feedback` function has been reworked to be recursive (it
shouldn't loop more than a couple times, and is tail recursive). This is
to extract the important part of the loop and be able to use it in
`oversee`, where the code was simply copy/pasted before.