Reimplement temporary drop management in Game
9e5c60a removed the ability to handle temporary drops in connectivity.
Now it is reimplemented in the `Game` setting, to avoid inconsistencies.

A `Game` is now in charge of remembering whether a client has lost
connection, and can act accordingly.

+ Improvement in error management: now the broadcast functions return a
`Disconnected` error when they encounter an invalid `id` they are told
to broadcast to. This leads to the game shutting down, as it is assumed
to be an unrecoverable error.