Fix the retry count being kept between games
When introducing the new creation page, a regression occured: the
count of attempts to reconnect to the server wasn't set properly anymore
to 0.

As a result, the game would refuse to start, even after creating a new
room. Which is not an intended or whishable behavior.

Now the reconnection attempts count is properly set to 0 when creating
and joining a new room.