Fix duration estimate failure at load
The uninitialized "categories" field caused the transformation of the
user data into sanitized data to fail. The duration estimate relied on
that transformation to work properly.

To estimate the game length, it isn't necessary to know certain things.
Now only the relevant data is validated to create a duration estimate.